+++ docs/release-notes.rst
@@ -7,6 +7,8 @@ History
 
 * Moved to https://github.com/codrsquad/runez
 
+* Corrected ``json_sanitized()``
+
 * Added ``@cached_property``, ``@chill_property``, ``is_basetype()``, ``is_iterable()``, ``joined()``
 
 * Added ``Struct`` schema type (for non-root serializable objects)